Create developing the solving part of the project
This part of the code is related to the part of solving relations and formulas of the main platform and this part will be updated in this direction. This part is still incomplete and since I have changed my code several times, I will share the current version with you, but this part of the code will definitely be changed and updated to complete the modeling part of the project.