Add files via upload
In this updated code, I worked on adding the optimization part for the prosumer decision model and also linking this part of the modeling with internal pricing by the coordinator agent. The main parts of the code remained the same as the original code, but the part related to optimization has been worked on. Due to the changes in this section and the change of the decision model compared to the previous code, the output results of the modeling have some errors compared to the previous code, which will be completed and corrected by modifying the changes compared to the previous code of this section.